Styling and adjustments
I created a function that handles the icons and their enabled status, which has fixed the problems with how it was handling loads from the list and adding a new load. Things now update properly.

I also reworked the QML a little to remove a warning it was spamming. The icons and things are now contained in a ColumnLayout and correctly respaced I believe.